Reference

API Reference

Policy graphs

Kokako.LinearPolicyGraph — Function.

LinearPolicyGraph(builder::Function; stages::Int, kwargs...)

Create a linear policy graph with stages number of stages.

See PolicyGraph for the other keyword arguments.

Kokako.MarkovianPolicyGraph — Function.

MarkovianPolicyGraph(builder::Function;
    transition_matrices::Vector{Array{Float64, 2}}, kwargs...)

Create a Markovian policy graph based on the transition matrices given in transition_matrices.

See PolicyGraph for the other keyword arguments.

Kokako.PolicyGraph — Type.

PolicyGraph(builder::Function, graph::Graph{T};
            bellman_function = AverageCut,
            optimizer = nothing,
            direct_mode = true) where T

Construct a a policy graph based on the graph structure of graph. (See Graph for details.)

Example

function builder(subproblem::JuMP.Model, index)
    # ... subproblem definition ...
end
model = PolicyGraph(builder, graph;
                    bellman_function = AverageCut,
                    optimizer = with_optimizer(GLPK.Optimizer),
                    direct_mode = false)

Or, using the Julia do ... end syntax:

model = PolicyGraph(graph;
                    bellman_function = AverageCut,
                    optimizer = with_optimizer(GLPK.Optimizer),
                    direct_mode = true) do subproblem, index
    # ... subproblem definitions ...
end

Subproblem definition

Kokako.@stageobjective — Macro.

@stageobjective(subproblem, expr)

Set the stage-objective of subproblem to expr.

Example

@stageobjective(subproblem, 2x + y)

Kokako.parameterize — Function.

parameterize(modify::Function,
             subproblem::JuMP.Model,
             realizations::Vector{T},
             probability::Vector{Float64} = fill(1.0 / length(realizations))
                 ) where T

Add a parameterization function modify to subproblem. The modify function takes one argument and modifies subproblem based on the realization of the noise sampled from realizations with corresponding probabilities probability.

In order to conduct an out-of-sample simulation, modify should accept arguments that are not in realizations (but still of type T).

Example

Kokako.parameterize(subproblem, [1, 2, 3], [0.4, 0.3, 0.3]) do ω
    JuMP.set_upper_bound(x, ω)
end

Training the policy

Kokako.train — Function.

Kokako.train(graph::PolicyGraph; kwargs...)::TrainingResults

Train the policy of the graph. Keyword arguments are

iterationlimit: number of iterations to conduct before termination. Defaults to 100000.

time_limit: number of seconds to train before termination. Defaults to Inf.

print_level: control the level of printing to the screen.

sampling_scheme: a sampling scheme to use on the forward pass of the algorithm. Defaults to InSampleMonteCarlo().

There is also a special option for infinite horizon problems

cyclediscretizationdelta: the maximum distance between states allowed on the forward pass.

Kokako.termination_status — Function.

termination_status(results::TrainingResults)

Query the reason why the training stopped.

Simulating the policy

Kokako.simulate — Function.

simulate(graph::PolicyGraph,
         number_replications::Int = 1,
         variables::Vector{Symbol} = Symbol[];
         sampling_scheme::AbstractSamplingScheme =
             InSampleMonteCarlo(),
         custom_recorders = Dict{Symbol, Function}()
 )::Vector{Vector{Dict{Symbol, Any}}}

Perform a simulation of the policy graph with number_replications replications using the sampling scheme sampling_scheme.

Returns a vector with one element for each replication. Each element is a vector with one-element for each node in the scenario that was sampled. Each element in that vector is a dictionary containing information about the subproblem that was solved.

In that dictionary there are four special keys:

:node_index, which records the index of the sampled node in the policy graph

:noise_term, which records the noise observed at the node

:stage_objective, which records the stage-objective of the subproblem

:bellman_term, which records the cost/value-to-go of the node.

The sum of :stageobjective + :bellmanterm will equal the objective value of the solved subproblem.

In addition to the special keys, the dictionary will contain the result of JuMP.value(subproblem[key]) for each key in variables. This is useful to obtain the primal value of the state and control variables.

For more complicated data, the custom_recorders keyword arguement can be used.

data = Dict{Symbol, Any}()
for (key, recorder) in custom_recorders
    data[key] = foo(subproblem)
end

For example, to record the dual of a constraint named my_constraint, pass the following:

simulation_results = simulate(graph, number_replications=2;
    custom_recorders = Dict(
        :constraint_dual = (sp) -> JuMP.dual(sp[:my_constraint])
    )
)

The value of the dual in the first stage of the second replication can be accessed as:

simulation_results[2][1][:constraint_dual]

Kokako.calculate_bound — Function.

Kokako.calculate_bound(graph::PolicyGraph, state::Dict{Symbol, Float64},
                       risk_measure=Expectation())

Calculate the lower bound (if minimizing, otherwise upper bound) of the problem graph at the point state, assuming the risk measure at the root node is risk_measure.

Kokako.Historical — Type.

Historical(scenarios::Vector{Vector{Tuple{T, S}}},
           probability::Vector{Float64})

A sampling scheme that samples a scenario from the vector of scenarios scenarios according to probability. If probability omitted, defaults to uniform probability.

Example

Historical(
    [
        [(1, 0.5), (2, 1.0), (3, 0.5)],
        [(1, 0.5), (2, 0.0), (3, 1.0)],
        [(1, 1.0), (2, 0.0), (3, 0.0)]
    ],
    [0.2, 0.5, 0.3]
)

Historical(scenario::Vector{Tuple{T, S}})

A deterministic sampling scheme that always samples scenario with probability 1.

Example

Historical([(1, 0.5), (2, 1.5), (3, 0.75)])